--- adventure.datetime/src/ca/cmput301f13t03/adventure_datetime/view/AuthorStoryDescriptionFragment.java
@@ -1,79 +1,0 @@
-/*
- *	Copyright (c) 2013 Andrew Fontaine, James Finlay, Jesse Tucker, Jacob Viau, and
- * 	Evan DeGraff
- *
- * 	Permission is hereby granted, free of charge, to any person obtaining a copy of
- * 	this software and associated documentation files (the "Software"), to deal in
- * 	the Software without restriction, including without limitation the rights to
- * 	use, copy, modify, merge, publish, distribute, sublicense, and/or sell copies of
- * 	the Software, and to permit persons to whom the Software is furnished to do so,
- * 	subject to the following conditions:
- *
- * 	The above copyright notice and this permission notice shall be included in all
- * 	copies or substantial portions of the Software.
- *
- * 	THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
- * 	IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS
- * 	FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR
- * 	COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER
- * 	IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN
- * 	CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
- */
-
-package ca.cmput301f13t03.adventure_datetime.view;
-
-import android.os.Bundle;
-import android.support.v4.app.Fragment;
-import android.util.Log;
-import android.view.LayoutInflater;
-import android.view.Menu;
-import android.view.MenuInflater;
-import android.view.MenuItem;
-import android.view.View;
-import android.view.ViewGroup;
-import ca.cmput301f13t03.adventure_datetime.R;
-
-public class AuthorStoryDescriptionFragment extends Fragment {
-	private static final String TAG = "AuthorStoryDescriptionFragment";
-	public static final String ARG_OBJECT = "object";
-
-	public void onCreate(Bundle bundle) {
-		super.onCreate(bundle);
-		setHasOptionsMenu(true);
-	}
-	
-	@Override
-	public View onCreateView(LayoutInflater inflater, ViewGroup container, Bundle savedInstanceState) {
-
-		View rootView = inflater.inflate(R.layout.listviewitem, container, false);
-		Bundle args = getArguments();
-
-		//TextView content = (TextView) rootView.findViewById(R.id.content);
-		//content.setText(Integer.toString(args.getInt(ARG_OBJECT)));
-
-		return rootView;
-	}
-
-	@Override
-	public void onCreateOptionsMenu(Menu menu, MenuInflater inflater) {
-		inflater.inflate(R.menu.authordesc, menu);
-	}
-
-	@Override
-	public boolean onOptionsItemSelected(MenuItem item) {
-
-		switch (item.getItemId()) {
-		case R.id.action_edit:
-			break;
-		case R.id.action_upload:
-			break;
-		case R.id.action_discard:
-			break;
-		default:
-			Log.e(TAG, "onOptionsItemSelected -> Unknown MenuItem");
-			break;
-		}
-		
-		return super.onOptionsItemSelected(item);
-	}
-}
